+++ Build1/zmachine.Library/Extensions/MachineOpcodeExtensions.VariableOperand.cs
@@ -111,13 +111,13 @@
 
         public static Machine processVAR(this Machine machine, int opcode, List<ushort> operands)
         {
-            if (!System.Enum.IsDefined(typeof(TwoOperandOpcode), (byte)opcode))
+            if (!System.Enum.IsDefined(typeof(TwoOperandOpcode), opcode))
             {
                 throw new InvalidCastException(opcode + " is not a defined value for enum type " +
                   typeof(TwoOperandOpcode).FullName);
             }
 
-            VariableOperandOpcode variableOperandOpcode = (VariableOperandOpcode)(byte)opcode;
+            VariableOperandOpcode variableOperandOpcode = (VariableOperandOpcode)opcode;
             switch (variableOperandOpcode)
             {
                 case VariableOperandOpcode.op_call:
